requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
python-nmap==0.7.1
cryptography==41.0.7
pyjwt==2.8.0
validators==0.22.0
pymongo==4.6.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
lxml==4.9.3
urllib3==2.1.0
python-dotenv==1.0.0
werkzeug==3.0.1
jinja2==3.1.2
reportlab==4.0.7
Pillow==10.1.0
passlib==1.7.4
bcrypt==4.1.1

